apps/api-shizen/requirements.txt
# SHINKOFA SHIZEN-PLANNER SERVICE - Python Dependencies
# Python: 3.11+

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM & Agents
langchain==0.1.0
langchain-community==0.0.10
ollama==0.1.6

# Vector Database & Embeddings
chromadb==0.4.18
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0

# Database
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.9  # Required for Alembic migrations
asyncpg>=0.31.0  # Async PostgreSQL driver
sqlalchemy>=2.0.45  # Python 3.13 compatibility
alembic==1.13.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing

# Payment Processing
stripe==11.1.1  # Stripe payments & subscriptions

# WebSocket
websockets==12.0

# File Operations
watchdog==3.0.0
python-dotenv==1.0.0

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# HTTP Client
httpx==0.25.2
requests>=2.32.3,<3.0.0  # Updated for kerykeion 4.14.1 compatibility

# Utilities
python-multipart==0.0.6
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
faker==20.1.0

# Linting & Formatting
ruff==0.1.6
mypy==1.7.1

# Type Stubs
types-requests==2.31.0.10
types-redis==4.6.0.11

# Monitoring & Logging
python-json-logger==2.0.7

# Design Humain & Astrology
pyswisseph==2.10.3.2  # Swiss Ephemeris for Design Humain calculations
kerykeion==4.14.1  # Astrology library (Western + Houses + Aspects)
